Refactor the hook
