+++ public/app/features/alerting/unified/utils/rules.ts
@@ -44,7 +44,7 @@ import { RULER_NOT_SUPPORTED_MSG } from './constants';
 import { getRulesSourceName, isGrafanaRulesSource } from './datasource';
 import { GRAFANA_ORIGIN_LABEL } from './labels';
 import { AsyncRequestState } from './redux';
-import { safeParsePrometheusDuration } from './time';
+import { formatPrometheusDuration, safeParsePrometheusDuration } from './time';
 
 export function isAlertingRule(rule: Rule | undefined): rule is AlertingRule {
   return typeof rule === 'object' && rule.type === PromRuleType.Alerting;
@@ -135,6 +135,26 @@ export function getRuleHealth(health: string): RuleHealth | undefined {
     default:
       return undefined;
   }
+}
+
+export function getPendingPeriod(rule: CombinedRule): string | undefined {
+  if (isRecordingRulerRule(rule.rulerRule) || isRecordingRule(rule.promRule)) {
+    return undefined;
+  }
+
+  // We prefer the for duration from the ruler rule because it is formatted as a duration string
+  // Prometheus duration is in seconds and we need to format it as a duration string
+  // Additionally, due to eventual consistency of the Prometheus endpoint the ruler data might be newer
+  if (isAlertingRulerRule(rule.rulerRule)) {
+    return rule.rulerRule.for;
+  }
+
+  if (isAlertingRule(rule.promRule)) {
+    const durationInMilliseconds = (rule.promRule.duration ?? 0) * 1000;
+    return formatPrometheusDuration(durationInMilliseconds);
+  }
+
+  return undefined;
 }
 
 export interface RulePluginOrigin {
